Update README.md
Data on its own has no value. It is the insights derived from the data that increases its value. These insights can be used to direct activities as varied as customer segmentation, demand and churn prediction, pricing optimization, retention marketing, and cost management - allowing businesses not to leave money on the table and to tap into new revenue opportunities.
Data monetization is not merely about the collection and storing of data; but extracting insights that can inform competitive advantages, differentiation and innovation strategies.